Prevent MediaScanner from removing Micro thumbnail database after scan.
After scanning the sdcard for the first time, the MediaScanner deletes
the micro thumbnail database files. The MiniThumbFile class attempts to
recreate the files but ends up with 0-sized db-files that it cannot
store any more micro thumbs in, resulting in that all micro thumbnails
that are results from thumbnail request are never stored properly until
the device is restarted.

+++ media/java/android/media/MediaScanner.java
@@ -303,7 +303,7 @@ public class MediaScanner
 
     // used when scanning the image database so we know whether we have to prune
     // old thumbnail files
-    private int mOriginalCount;
+    private int mOriginalImageCount;
     /** Whether the scanner has set a default sound for the ringer ringtone. */
     private boolean mDefaultRingtoneSet;
     /** Whether the scanner has set a default sound for the notification ringtone. */
@@ -892,6 +892,7 @@ public class MediaScanner
         Cursor c = null;
         String where = null;
         String[] selectionArgs = null;
+        int originalVideoCount = 0;
 
         if (mFileCache == null) {
             mFileCache = new HashMap<String, FileCacheEntry>();
@@ -943,6 +944,7 @@ public class MediaScanner
 
             if (c != null) {
                 try {
+                    originalVideoCount = c.getCount();
                     while (c.moveToNext()) {
                         long rowId = c.getLong(ID_VIDEO_COLUMN_INDEX);
                         String path = c.getString(PATH_VIDEO_COLUMN_INDEX);
@@ -967,12 +969,12 @@ public class MediaScanner
             } else {
                 where = null;
             }
-            mOriginalCount = 0;
+            mOriginalImageCount = 0;
             c = mMediaProvider.query(mImagesUri, IMAGES_PROJECTION, where, selectionArgs, null);
 
             if (c != null) {
                 try {
-                    mOriginalCount = c.getCount();
+                    mOriginalImageCount = c.getCount();
                     while (c.moveToNext()) {
                         long rowId = c.getLong(ID_IMAGES_COLUMN_INDEX);
                         String path = c.getString(PATH_IMAGES_COLUMN_INDEX);
@@ -1029,6 +1031,17 @@ public class MediaScanner
                 c.close();
             }
         }
+
+        // Clear micro thumb databases now if applicable, instead of after scan
+        // to avoid undoing work that has been done by the thumbnail queue
+        // during scan.
+        if (originalVideoCount == 0 && mImagesUri.equals(Video.Media.getContentUri("external")))
+            MiniThumbFile.instance(MediaStore.Video.Media.EXTERNAL_CONTENT_URI)
+                    .removeMiniThumbDataFile();
+
+        if (mOriginalImageCount == 0 && mImagesUri.equals(Images.Media.getContentUri("external")))
+            MiniThumbFile.instance(MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
+                    .removeMiniThumbDataFile();
     }
 
     private boolean inScanDirectory(String path, String[] directories) {
@@ -1071,7 +1084,10 @@ public class MediaScanner
                 if (Config.LOGV)
                     Log.v(TAG, "fileToDelete is " + fileToDelete);
                 try {
-                    (new File(fileToDelete)).delete();
+                    // Exclude micro thumb data files that have already been handled in prescan
+                    if (!fileToDelete.contains(".thumbdata")) {
+                        (new File(fileToDelete)).delete();
+                    }
                 } catch (SecurityException ex) {
                 }
             }
@@ -1133,7 +1149,7 @@ public class MediaScanner
             processPlayLists();
         }
 
-        if (mOriginalCount == 0 && mImagesUri.equals(Images.Media.getContentUri("external")))
+        if (mOriginalImageCount == 0 && mImagesUri.equals(Images.Media.getContentUri("external")))
             pruneDeadThumbnailFiles();
 
         // allow GC to clean up
